requirements.txt
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
pyzmq==16.0.2
Werkzeug==0.11.15
